+++ eras/babbage/impl/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.6.0.0
 
+* Add `ToJSON` for `BabbageContextError`
 * `FromJSON` instance for `PParamsUPdate`: #3949
 * Fixed JSON serialization of `PParams` to include the `protocolVersion`: #3953
 * Stop exporting all of the internal `hkd*` functions and `PParamsHKD` from